Implemented initial UI setup for Routing (#24)
* [dist]: Added ionic and react router dependencies.

Added ionic and react router dependencies. For UI and routing.

* [feat]: Updated layout for UI and routing.

Updated layout for ionic to run and render UI properly and with the routing layout done.

* [feat]: Resources.

Made resources folder and added banner image.

* [feat]: Started homepage and header.

Implemented basic homepage setup and added the header for it.

* [feat]: Basic navbar and browse page.

Implemented basic navbar with the colour scheme on browse page.

* [refactor]: Browse to Food Details.

Changed Browse page to Food Details page.

* [feat]: Added new pages.

Added new pages to be routed to.

* [feat]: Not Found.

Not Found page routing and incorrect link redirection to it implemented.