Fix food card props (#87)
* [dist]: Added ionic and react router dependencies.

Added ionic and react router dependencies. For UI and routing.

* [feat]: Updated layout for UI and routing.

Updated layout for ionic to run and render UI properly and with the routing layout done.

* [feat]: Resources.

Made resources folder and added banner image.

* [feat]: Started homepage and header.

Implemented basic homepage setup and added the header for it.

* [feat]: Basic navbar and browse page.

Implemented basic navbar with the colour scheme on browse page.

* [refactor]: Browse to Food Details.

Changed Browse page to Food Details page.

* [feat]: Added new pages.

Added new pages to be routed to.

* [feat]: Not Found.

Not Found page routing and incorrect link redirection to it implemented.

* [style]: Added ionic styles.

Added ionic styles to index.css.

* [style]: Updated themes.

Updated themes to account for different texts and backgrounds - dark and light.

* [fix]: Search bar text colour.

Search bar text colour needed to be specified explicitly.

* [feat]: Post button and ion style.

Post button main structure implemented.
New ion style implemented.

* [feat]: Finalised AddPostButton.

Added the main features of  the AddPostButton, just the add post modal needs to be added in the commented marker when it is implemented.

* [feat]: Posts Layout component.

Posts Layout component implemented.
Need to sort out how to pass the cards for each.

* [fix]: FoodCards generated based on input props.

Fixed mapping. Props have data and FoodCards are made using those, FoodCards are not the props themselves.

Will need to update later as prop details get updated with FoodCard props.

* [fix]: Default value.

Default value of props set to empty array.

* [feat]: Updated FoodCard.

Updated FoodCard to include props. Like button can be clicked to change whether it is liked or not. Image can be clicked to go to the relevant details page (need backend so currently implemented with homepage).

* Delete Navbar.js

* [refactor]: FoodCard refactored.

Extra function removed and functionality integrated with FoodCard component.

* [feat]: Updating number of likes in UI.

Updating number of likes in UI. Still need to accommodate for backend functionality.

Co-authored-by: Yujia Wu <49899381+ywu666@users.noreply.github.com>